# backend/src/main/java/com/diceduel/dto/CreatePlayerRequest.java
package com.diceduel.dto;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;

public record CreatePlayerRequest(
        @NotBlank @Size(max = 80) String name
) {
}